add vhdl component for sigmoid and tanh
